+++ src/Executor/ReferenceExecutor.php
@@ -451,9 +451,9 @@ class ReferenceExecutor implements ExecutorImplementation
     protected function shouldIncludeNode(SelectionNode $node): bool
     {
         $variableValues = $this->exeContext->variableValues;
-        $skipDirective  = Directive::skipDirective();
-        $skip           = Values::getDirectiveValues(
-            $skipDirective,
+
+        $skip = Values::getDirectiveValues(
+            Directive::skipDirective(),
             $node,
             $variableValues
         );
@@ -461,9 +461,8 @@ class ReferenceExecutor implements ExecutorImplementation
             return false;
         }
 
-        $includeDirective = Directive::includeDirective();
-        $include          = Values::getDirectiveValues(
-            $includeDirective,
+        $include = Values::getDirectiveValues(
+            Directive::includeDirective(),
             $node,
             $variableValues
         );
@@ -830,11 +829,7 @@ class ReferenceExecutor implements ExecutorImplementation
         if ($returnType !== $this->exeContext->schema->getType($returnType->name)) {
             $hint = '';
             if ($this->exeContext->schema->getConfig()->typeLoader !== null) {
-                $hint = sprintf(
-                    'Make sure that type loader returns the same instance as defined in %s.%s',
-                    $info->parentType,
-                    $info->fieldName
-                );
+                $hint = "Make sure that type loader returns the same instance as defined in {$info->parentType}.{$info->fieldName}";
             }
 
             throw new InvariantViolation(