+++ src/Type/Definition/InputObjectField.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace GraphQL\Type\Definition;
 
-use GraphQL\Error\Error;
 use GraphQL\Error\InvariantViolation;
 use GraphQL\Language\AST\InputValueDefinitionNode;
 use GraphQL\Type\Schema;
@@ -51,12 +50,10 @@ class InputObjectField
     {
         if (! isset($this->type)) {
             /**
-             * TODO: replace this phpstan cast with native assert
-             *
-             * @var Type&InputType
+             * @see it('rejects an Input Object type with incorrectly typed fields')
              */
-            $type       = Schema::resolveType($this->config['type']);
-            $this->type = $type;
+             // @phpstan-ignore-next-line schema validation will catch a Type that is not an InputType
+            $this->type = Schema::resolveType($this->config['type']);
         }
 
         return $this->type;
@@ -78,10 +75,9 @@ class InputObjectField
      */
     public function assertValid(Type $parentType): void
     {
-        try {
-            Utils::assertValidName($this->name);
-        } catch (Error $e) {
-            throw new InvariantViolation(sprintf('%s.%s: %s', $parentType->name, $this->name, $e->getMessage()));
+        $error = Utils::isValidNameError($this->name);
+        if ($error !== null) {
+            throw new InvariantViolation("{$parentType->name}.{$this->name}: {$error->getMessage()}");
         }
 
         $type = $this->getType();