+++ src/Type/SchemaValidationContext.php
@@ -9,7 +9,6 @@ use GraphQL\Language\AST\DirectiveDefinitionNode;
 use GraphQL\Language\AST\DirectiveNode;
 use GraphQL\Language\AST\EnumTypeDefinitionNode;
 use GraphQL\Language\AST\EnumTypeExtensionNode;
-use GraphQL\Language\AST\EnumValueDefinitionNode;
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\InputObjectTypeDefinitionNode;
 use GraphQL\Language\AST\InputObjectTypeExtensionNode;
@@ -23,12 +22,8 @@ use GraphQL\Language\AST\NodeList;
 use GraphQL\Language\AST\NonNullTypeNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeExtensionNode;
-use GraphQL\Language\AST\OperationTypeDefinitionNode;
-use GraphQL\Language\AST\ScalarTypeDefinitionNode;
-use GraphQL\Language\AST\ScalarTypeExtensionNode;
 use GraphQL\Language\AST\SchemaDefinitionNode;
 use GraphQL\Language\AST\SchemaTypeExtensionNode;
-use GraphQL\Language\AST\TypeDefinitionNode;
 use GraphQL\Language\AST\TypeNode;
 use GraphQL\Language\AST\UnionTypeDefinitionNode;
 use GraphQL\Language\AST\UnionTypeExtensionNode;
@@ -57,6 +52,7 @@ use function count;
 use function in_array;
 use function is_array;
 use function is_object;
+use function property_exists;
 use function sprintf;
 
 class SchemaValidationContext
@@ -110,31 +106,6 @@ class SchemaValidationContext
         $this->errors[] = $error;
     }
 
-    /**
-     * @return NamedTypeNode|(Node&TypeDefinitionNode)|null
-     */
-    private function getOperationTypeNode(Type $type, string $operation): ?Node
-    {
-        $astNode = $this->schema->getAstNode();
-
-        $operationTypeNode = null;
-        if ($astNode instanceof SchemaDefinitionNode) {
-            /** @var OperationTypeDefinitionNode|null $operationTypeNode */
-            $operationTypeNode = null;
-
-            foreach ($astNode->operationTypes as $operationType) {
-                if ($operationType->operation === $operation) {
-                    $operationTypeNode = $operationType;
-                    break;
-                }
-            }
-        }
-
-        return $operationTypeNode === null
-            ? ($type === null ? null : $type->astNode)
-            : $operationTypeNode->type;
-    }
-
     public function validateDirectives(): void
     {
         $this->validateDirectiveDefinitions();
@@ -153,14 +124,17 @@ class SchemaValidationContext
         $directives = $this->schema->getDirectives();
         foreach ($directives as $directive) {
             // Ensure all directives are in fact GraphQL directives.
+            // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
             if (! $directive instanceof Directive) {
-                $nodes = is_object($directive)
+                $notDirective = Utils::printSafe($directive);
+                // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
+                $node = is_object($directive) && property_exists($directive, 'astNode')
                     ? $directive->astNode
                     : null;
 
                 $this->reportError(
-                    'Expected directive but got: ' . Utils::printSafe($directive) . '.',
-                    $nodes
+                    "Expected directive but got: {$notDirective}.",
+                    $node
                 );
                 continue;
             }
@@ -183,7 +157,7 @@ class SchemaValidationContext
 
                 if (isset($argNames[$argName])) {
                     $this->reportError(
-                        sprintf('Argument @%s(%s:) can only be defined once.', $directive->name, $argName),
+                        "Argument @{$directive->name}({$argName}:) can only be defined once.",
                         $this->getAllDirectiveArgNodes($directive, $argName)
                     );
                     continue;
@@ -196,13 +170,9 @@ class SchemaValidationContext
                     continue;
                 }
 
+                $type = Utils::printSafe($arg->getType());
                 $this->reportError(
-                    sprintf(
-                        'The type of @%s(%s:) must be Input Type but got: %s.',
-                        $directive->name,
-                        $argName,
-                        Utils::printSafe($arg->getType())
-                    ),
+                    "The type of @{$directive->name}({$argName}:) must be Input Type but got: {$type}.",
                     $this->getDirectiveArgTypeNode($directive, $argName)
                 );
             }
@@ -213,18 +183,18 @@ class SchemaValidationContext
                 continue;
             }
 
-            $nodes = [];
+            $node = [];
             foreach ($directiveList as $dir) {
                 if ($dir->astNode === null) {
                     continue;
                 }
 
-                $nodes[] = $dir->astNode;
+                $node[] = $dir->astNode;
             }
 
             $this->reportError(
                 sprintf('Directive @%s defined multiple times.', $directiveName),
-                $nodes
+                $node
             );
         }
     }
@@ -251,22 +221,21 @@ class SchemaValidationContext
      */
     private function getAllDirectiveArgNodes(Directive $directive, string $argName): array
     {
-        $subNodes = $this->getAllSubNodes(
-            $directive,
-            /**
-             * @return NodeList<InputValueDefinitionNode>
-             */
-            static function (DirectiveDefinitionNode $directiveNode): NodeList {
-                return $directiveNode->arguments;
-            }
-        );
+        $astNode = $directive->astNode;
+        if ($astNode === null) {
+            return [];
+        }
 
-        return Utils::filter(
-            $subNodes,
-            static function (InputValueDefinitionNode $argNode) use ($argName): bool {
-                return $argNode->name->value === $argName;
+        $matchingSubnodes = [];
+        foreach ($astNode->arguments as $subNode) {
+            if ($subNode->name->value !== $argName) {
+                continue;
             }
-        );
+
+            $matchingSubnodes[] = $subNode;
+        }
+
+        return $matchingSubnodes;
     }
 
     /**
@@ -284,12 +253,19 @@ class SchemaValidationContext
     public function validateTypes(): void
     {
         $typeMap = $this->schema->getTypeMap();
-        foreach ($typeMap as $typeName => $type) {
+        foreach ($typeMap as $type) {
             // Ensure all provided types are in fact GraphQL type.
+            // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
             if (! $type instanceof NamedType) {
+                $notNamedType = Utils::printSafe($type);
+                // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
+                $node = $type instanceof Type
+                    ? $type->astNode
+                    : null;
+
                 $this->reportError(
-                    'Expected GraphQL named type but got: ' . Utils::printSafe($type) . '.',
-                    $type instanceof Type ? $type->astNode : null
+                    "Expected GraphQL named type but got: {$notNamedType}.",
+                    $node
                 );
                 continue;
             }
@@ -428,7 +404,7 @@ class SchemaValidationContext
         // Objects and Interfaces both must define one or more fields.
         if ($fieldMap === []) {
             $this->reportError(
-                sprintf('Type %s must define one or more fields.', $type->name),
+                "Type {$type->name} must define one or more fields.",
                 $this->getAllNodes($type)
             );
         }
@@ -543,49 +519,29 @@ class SchemaValidationContext
     }
 
     /**
-     * @param Schema|ObjectType|InterfaceType|UnionType|EnumType|Directive $obj
-     * @param callable(Node): (iterable<Node>|null)                        $getter
-     */
-    private function getAllSubNodes(object $obj, callable $getter): NodeList
-    {
-        $result = new NodeList([]);
-        foreach ($this->getAllNodes($obj) as $astNode) {
-            if ($astNode === null) {
-                continue;
-            }
-
-            $subNodes = $getter($astNode);
-            if ($subNodes === null) {
-                continue;
-            }
-
-            $result = $result->merge($subNodes);
-        }
-
-        return $result;
-    }
-
-    /**
      * @param ObjectType|InterfaceType $type
      *
      * @return array<int, FieldDefinitionNode>
      */
     private function getAllFieldNodes(Type $type, string $fieldName): array
     {
-        $subNodes = $this->getAllSubNodes(
-            $type,
-            /**
-             * @return NodeList<FieldDefinitionNode>
-             */
-            static function (Node $typeNode): NodeList {
-                /** @var ObjectTypeDefinitionNode|ObjectTypeExtensionNode|InterfaceTypeDefinitionNode|InterfaceTypeExtensionNode $typeNode */
-                return $typeNode->fields;
-            }
-        );
+        $allNodes = $type->astNode !== null
+            ? array_merge([$type->astNode], $type->extensionASTNodes)
+            : $type->extensionASTNodes;
 
-        return Utils::filter($subNodes, static function ($fieldNode) use ($fieldName): bool {
-            return $fieldNode->name->value === $fieldName;
-        });
+        $matchingFieldNodes = [];
+
+        foreach ($allNodes as $node) {
+            foreach ($node->fields as $field) {
+                if ($field->name->value !== $fieldName) {
+                    continue;
+                }
+
+                $matchingFieldNodes[] = $field;
+            }
+        }
+
+        return $matchingFieldNodes;
     }
 
     /**
@@ -664,36 +620,37 @@ class SchemaValidationContext
     private function validateInterfaces(ImplementingType $type): void
     {
         $ifaceTypeNames = [];
-        foreach ($type->getInterfaces() as $iface) {
-            if (! $iface instanceof InterfaceType) {
-                $safeIface = Utils::printSafe($iface);
+        foreach ($type->getInterfaces() as $interface) {
+            // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
+            if (! $interface instanceof InterfaceType) {
+                $notInterface = Utils::printSafe($interface);
                 $this->reportError(
-                    "Type {$type->name} must only implement Interface types, it cannot implement {$safeIface}.",
-                    $this->getImplementsInterfaceNode($type, $iface)
+                    "Type {$type->name} must only implement Interface types, it cannot implement {$notInterface}.",
+                    $this->getImplementsInterfaceNode($type, $interface)
                 );
                 continue;
             }
 
-            if ($type === $iface) {
+            if ($type === $interface) {
                 $this->reportError(
                     "Type {$type->name} cannot implement itself because it would create a circular reference.",
-                    $this->getImplementsInterfaceNode($type, $iface)
+                    $this->getImplementsInterfaceNode($type, $interface)
                 );
                 continue;
             }
 
-            if (isset($ifaceTypeNames[$iface->name])) {
+            if (isset($ifaceTypeNames[$interface->name])) {
                 $this->reportError(
-                    "Type {$type->name} can only implement {$iface->name} once.",
-                    $this->getAllImplementsInterfaceNodes($type, $iface)
+                    "Type {$type->name} can only implement {$interface->name} once.",
+                    $this->getAllImplementsInterfaceNodes($type, $interface)
                 );
                 continue;
             }
 
-            $ifaceTypeNames[$iface->name] = true;
+            $ifaceTypeNames[$interface->name] = true;
 
-            $this->validateTypeImplementsAncestors($type, $iface);
-            $this->validateTypeImplementsInterface($type, $iface);
+            $this->validateTypeImplementsAncestors($type, $interface);
+            $this->validateTypeImplementsInterface($type, $interface);
         }
     }
 
@@ -704,16 +661,14 @@ class SchemaValidationContext
      */
     private function getDirectives(object $object): NodeList
     {
-        return $this->getAllSubNodes(
-            $object,
-            /**
-             * @return NodeList<DirectiveNode>
-             */
-            static function (Node $node): NodeList {
-                /** @var SchemaDefinitionNode|SchemaTypeExtensionNode|ObjectTypeDefinitionNode|ObjectTypeExtensionNode|InterfaceTypeDefinitionNode|InterfaceTypeExtensionNode|UnionTypeDefinitionNode|UnionTypeExtensionNode|EnumTypeDefinitionNode|EnumTypeExtensionNode|InputObjectTypeDefinitionNode|InputObjectTypeExtensionNode|ScalarTypeDefinitionNode|ScalarTypeExtensionNode $node */
-                return $node->directives;
+        $directives = [];
+        foreach ($this->getAllNodes($object) as $node) {
+            foreach ($node->directives as $directive) {
+                $directives[] = $directive;
             }
-        );
+        }
+
+        return new NodeList($directives);
     }
 
     /**
@@ -733,20 +688,24 @@ class SchemaValidationContext
      */
     private function getAllImplementsInterfaceNodes(ImplementingType $type, Type $shouldBeInterface): array
     {
-        $subNodes = $this->getAllSubNodes(
-            $type,
-            /**
-             * @return NodeList<NamedTypeNode>
-             */
-            static function (Node $typeNode): NodeList {
-                /** @var ObjectTypeDefinitionNode|ObjectTypeExtensionNode|InterfaceTypeDefinitionNode|InterfaceTypeExtensionNode $typeNode */
-                return $typeNode->interfaces;
-            }
-        );
+        $allNodes = $type->astNode !== null
+            ? array_merge([$type->astNode], $type->extensionASTNodes)
+            : $type->extensionASTNodes;
 
-        return Utils::filter($subNodes, static function (NamedTypeNode $ifaceNode) use ($shouldBeInterface): bool {
-            return $ifaceNode->name->value === $shouldBeInterface->name;
-        });
+        $shouldBeInterfaceName  = $shouldBeInterface->name;
+        $matchingInterfaceNodes = [];
+
+        foreach ($allNodes as $node) {
+            foreach ($node->interfaces as $interface) {
+                if ($interface->name->value !== $shouldBeInterfaceName) {
+                    continue;
+                }
+
+                $matchingInterfaceNodes[] = $interface;
+            }
+        }
+
+        return $matchingInterfaceNodes;
     }
 
     /**
@@ -766,12 +725,7 @@ class SchemaValidationContext
             // Assert interface field exists on type.
             if ($typeField === null) {
                 $this->reportError(
-                    sprintf(
-                        'Interface field %s.%s expected but %s does not provide it.',
-                        $iface->name,
-                        $fieldName,
-                        $type->name
-                    ),
+                    "Interface field {$iface->name}.{$fieldName} expected but {$type->name} does not provide it.",
                     array_merge(
                         [$this->getFieldNode($iface, $fieldName)],
                         $this->getAllNodes($type)
@@ -936,7 +890,7 @@ class SchemaValidationContext
 
         if ($memberTypes === []) {
             $this->reportError(
-                sprintf('Union type %s must define one or more member types.', $union->name),
+                "Union type {$union->name} must define one or more member types.",
                 $this->getAllNodes($union)
             );
         }
@@ -944,14 +898,12 @@ class SchemaValidationContext
         $includedTypeNames = [];
 
         foreach ($memberTypes as $memberType) {
+            // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
             if (! $memberType instanceof ObjectType) {
+                $notObjectType = Utils::printSafe($memberType);
                 $this->reportError(
-                    sprintf(
-                        'Union type %s can only include Object types, it cannot include %s.',
-                        $union->name,
-                        Utils::printSafe($memberType)
-                    ),
-                    $this->getUnionMemberTypeNodes($union, Utils::printSafe($memberType))
+                    "Union type {$union->name} can only include Object types, it cannot include {$notObjectType}.",
+                    $this->getUnionMemberTypeNodes($union, $notObjectType)
                 );
                 continue;
             }
@@ -973,20 +925,22 @@ class SchemaValidationContext
      */
     private function getUnionMemberTypeNodes(UnionType $union, string $typeName): array
     {
-        $subNodes = $this->getAllSubNodes(
-            $union,
-            /**
-             * @return NodeList<NamedTypeNode>
-             */
-            static function (Node $unionNode): NodeList {
-                /** @var UnionTypeDefinitionNode|UnionTypeExtensionNode $unionNode */
-                return $unionNode->types;
-            }
-        );
+        $allNodes = $union->astNode !== null
+            ? array_merge([$union->astNode], $union->extensionASTNodes)
+            : $union->extensionASTNodes;
 
-        return Utils::filter($subNodes, static function (NamedTypeNode $typeNode) use ($typeName): bool {
-            return $typeNode->name->value === $typeName;
-        });
+        $types = [];
+        foreach ($allNodes as $node) {
+            foreach ($node->types as $type) {
+                if ($type->name->value !== $typeName) {
+                    continue;
+                }
+
+                $types[] = $type;
+            }
+        }
+
+        return $types;
     }
 
     private function validateEnumValues(EnumType $enumType): void
@@ -995,22 +949,13 @@ class SchemaValidationContext
 
         if ($enumValues === []) {
             $this->reportError(
-                sprintf('Enum type %s must define one or more values.', $enumType->name),
+                "Enum type {$enumType->name} must define one or more values.",
                 $this->getAllNodes($enumType)
             );
         }
 
         foreach ($enumValues as $enumValue) {
             $valueName = $enumValue->name;
-
-            // Ensure no duplicates
-            $allNodes = $this->getEnumValueNodes($enumType, $valueName);
-            if (count($allNodes) > 1) {
-                $this->reportError(
-                    sprintf('Enum type %s can include value %s only once.', $enumType->name, $valueName),
-                    $allNodes
-                );
-            }
 
             // Ensure valid name.
             $this->validateName($enumValue);
@@ -1033,35 +978,13 @@ class SchemaValidationContext
         }
     }
 
-    /**
-     * @return array<int, EnumValueDefinitionNode>
-     */
-    private function getEnumValueNodes(EnumType $enum, string $valueName): array
-    {
-        $subNodes = $this->getAllSubNodes(
-            $enum,
-            /**
-             * @param EnumTypeDefinitionNode|EnumTypeExtensionNode $enumNode
-             *
-             * @return NodeList<EnumValueDefinitionNode>
-             */
-            static function (Node $enumNode): NodeList {
-                return $enumNode->values;
-            }
-        );
-
-        return Utils::filter($subNodes, static function ($valueNode) use ($valueName): bool {
-            return $valueNode->name->value === $valueName;
-        });
-    }
-
     private function validateInputFields(InputObjectType $inputObj): void
     {
         $fieldMap = $inputObj->getFields();
 
         if ($fieldMap === []) {
             $this->reportError(
-                sprintf('Input Object type %s must define one or more fields.', $inputObj->name),
+                "Input Object type {$inputObj->name} must define one or more fields.",
                 $this->getAllNodes($inputObj)
             );
         }
@@ -1073,18 +996,14 @@ class SchemaValidationContext
 
             // TODO: Ensure they are unique per field.
 
-            // Ensure the type is an input type
-            if (! Type::isInputType($field->getType())) {
+            // Ensure the type is an input type.
+            $type = $field->getType();
+            // @phpstan-ignore-next-line The generic type says this should not happen, but a user may use it wrong nonetheless
+            if (! Type::isInputType($type)) {
+                $notInputType = Utils::printSafe($type);
                 $this->reportError(
-                    sprintf(
-                        'The type of %s.%s must be Input Type but got: %s.',
-                        $inputObj->name,
-                        $fieldName,
-                        Utils::printSafe($field->getType())
-                    ),
-                    $field->astNode !== null
-                        ? $field->astNode->type
-                        : null
+                    "The type of {$inputObj->name}.{$fieldName} must be Input Type but got: {$notInputType}.",
+                    $field->astNode->type ?? null
                 );
             }