+++ tests/Server/RequestParsingTest.php
@@ -10,6 +10,7 @@ use GraphQL\Server\OperationParams;
 use GraphQL\Server\RequestError;
 use InvalidArgumentException;
 use Nyholm\Psr7\Request;
+use Nyholm\Psr7\ServerRequest;
 use Nyholm\Psr7\Stream;
 use Nyholm\Psr7\Uri;
 use PHPUnit\Framework\TestCase;
@@ -72,23 +73,18 @@ class RequestParsingTest extends TestCase
     }
 
     /**
-     * @param OperationParams $params
-     * @param string          $query
-     * @param string          $queryId
-     * @param mixed|null      $variables
-     * @param string          $operation
+     * @param mixed                $variables
+     * @param array<string, mixed> $extensions
      */
     private static function assertValidOperationParams(
-        $params,
-        $query,
-        $queryId = null,
+        OperationParams $params,
+        ?string $query,
+        ?string $queryId = null,
         $variables = null,
-        $operation = null,
-        $extensions = null,
-        $message = ''
+        ?string $operation = null,
+        ?array $extensions = null,
+        string $message = ''
     ): void {
-        self::assertInstanceOf(OperationParams::class, $params, $message);
-
         self::assertSame($query, $params->query, $message);
         self::assertSame($queryId, $params->queryId, $message);
         self::assertSame($variables, $params->variables, $message);
@@ -110,6 +106,8 @@ class RequestParsingTest extends TestCase
         $parsed = [
             'raw' => $this->parseRawFormUrlencodedRequest($post),
             'psr' => $this->parsePsrFormUrlEncodedRequest($post),
+            'serverRequest' => $this->parsePsrFormUrlEncodedServerRequest($post, false),
+            'parsedServerRequest' => $this->parsePsrFormUrlEncodedServerRequest($post, true),
         ];
 
         foreach ($parsed as $method => $parsedBody) {
@@ -153,6 +151,24 @@ class RequestParsingTest extends TestCase
                 http_build_query($postValue)
             )
         );
+    }
+
+    private function parsePsrFormUrlEncodedServerRequest($postValue, bool $parsed)
+    {
+        $helper = new Helper();
+
+        $request = new ServerRequest(
+            'POST',
+            '',
+            ['Content-Type' => 'application/x-www-form-urlencoded'],
+            $parsed ? null : http_build_query($postValue),
+        );
+
+        if ($parsed) {
+            $request = $request->withParsedBody($postValue);
+        }
+
+        return $helper->parsePsrRequest($request);
     }
 
     public function testParsesGetRequest(): void