+++ tests/Type/DefinitionTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace GraphQL\Tests\Type;
 
 use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
+use Generator;
 use GraphQL\Error\InvariantViolation;
 use GraphQL\Tests\Type\TestClasses\MyCustomType;
 use GraphQL\Tests\Type\TestClasses\OtherCustom;
@@ -14,9 +15,11 @@ use GraphQL\Type\Definition\FieldDefinition;
 use GraphQL\Type\Definition\InputObjectField;
 use GraphQL\Type\Definition\InputObjectType;
 use GraphQL\Type\Definition\InterfaceType;
+use GraphQL\Type\Definition\IntType;
 use GraphQL\Type\Definition\ListOfType;
 use GraphQL\Type\Definition\NonNull;
 use GraphQL\Type\Definition\ObjectType;
+use GraphQL\Type\Definition\StringType;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Definition\UnionType;
 use GraphQL\Type\Schema;
@@ -164,35 +167,30 @@ class DefinitionTest extends TestCase
         self::assertSame($blogSchema->getQueryType(), $this->blogQuery);
 
         $articleField = $this->blogQuery->getField('article');
-        self::assertSame($articleField->getType(), $this->blogArticle);
-        self::assertSame($articleField->getType()->name, 'Article');
         self::assertSame($articleField->name, 'article');
 
-        /** @var ObjectType $articleFieldType */
         $articleFieldType = $articleField->getType();
-        $titleField       = $articleFieldType->getField('title');
+        self::assertInstanceOf(ObjectType::class, $articleFieldType);
+        self::assertSame($articleFieldType, $this->blogArticle);
+        self::assertSame($articleFieldType->name, 'Article');
 
-        self::assertInstanceOf('GraphQL\Type\Definition\FieldDefinition', $titleField);
+        $titleField = $articleFieldType->getField('title');
         self::assertSame('title', $titleField->name);
         self::assertSame(Type::string(), $titleField->getType());
 
         $authorField = $articleFieldType->getField('author');
-        self::assertInstanceOf('GraphQL\Type\Definition\FieldDefinition', $authorField);
 
-        /** @var ObjectType $authorFieldType */
         $authorFieldType = $authorField->getType();
+        self::assertInstanceOf(ObjectType::class, $authorFieldType);
         self::assertSame($this->blogAuthor, $authorFieldType);
 
         $recentArticleField = $authorFieldType->getField('recentArticle');
-        self::assertInstanceOf('GraphQL\Type\Definition\FieldDefinition', $recentArticleField);
         self::assertSame($this->blogArticle, $recentArticleField->getType());
 
         $feedField = $this->blogQuery->getField('feed');
-        self::assertInstanceOf('GraphQL\Type\Definition\FieldDefinition', $feedField);
 
-        /** @var ListOfType $feedFieldType */
         $feedFieldType = $feedField->getType();
-        self::assertInstanceOf('GraphQL\Type\Definition\ListOfType', $feedFieldType);
+        self::assertInstanceOf(ListOfType::class, $feedFieldType);
         self::assertSame($this->blogArticle, $feedFieldType->getWrappedType());
     }
 
@@ -207,12 +205,13 @@ class DefinitionTest extends TestCase
         ]);
 
         self::assertSame($this->blogMutation, $schema->getMutationType());
-        $writeMutation = $this->blogMutation->getField('writeArticle');
 
-        self::assertInstanceOf('GraphQL\Type\Definition\FieldDefinition', $writeMutation);
-        self::assertSame($this->blogArticle, $writeMutation->getType());
-        self::assertSame('Article', $writeMutation->getType()->name);
+        $writeMutation = $this->blogMutation->getField('writeArticle');
         self::assertSame('writeArticle', $writeMutation->name);
+
+        $writeMutationType = $writeMutation->getType();
+        self::assertSame($this->blogArticle, $writeMutationType);
+        self::assertSame('Article', $writeMutationType->name);
     }
 
     /**
@@ -553,6 +552,7 @@ class DefinitionTest extends TestCase
             ],
         ]);
 
+        /** @var ObjectType|null $blog */
         $blog   = null;
         $called = false;
 
@@ -601,14 +601,12 @@ class DefinitionTest extends TestCase
         self::assertSame([$node], $blog->getInterfaces());
         self::assertSame([$node], $user->getInterfaces());
 
-        self::assertNotNull($user->getField('blogs'));
-        /** @var NonNull $blogFieldReturnType */
         $blogFieldReturnType = $user->getField('blogs')->getType();
+        self::assertInstanceOf(NonNull::class, $blogFieldReturnType);
         self::assertSame($blog, $blogFieldReturnType->getWrappedType(true));
 
-        self::assertNotNull($blog->getField('owner'));
-        /** @var NonNull $ownerFieldReturnType */
         $ownerFieldReturnType = $blog->getField('owner')->getType();
+        self::assertInstanceOf(NonNull::class, $ownerFieldReturnType);
         self::assertSame($user, $ownerFieldReturnType->getWrappedType(true));
     }
 
@@ -1611,7 +1609,6 @@ class DefinitionTest extends TestCase
         self::assertInstanceOf(InputObjectField::class, $inputObjType->findField($fieldName));
         self::assertNull($inputObjType->findField('unknown'));
 
-        self::assertInstanceOf(InputObjectField::class, $inputObjType->getField($fieldName));
         self::expectExceptionObject(new InvariantViolation('Field "unknown" is not defined for type "SomeInputObject"'));
         $inputObjType->getField('unknown');
     }
@@ -1718,10 +1715,10 @@ class DefinitionTest extends TestCase
             'name'   => 'SomeEnum',
             'values' => [['FOO' => 10]],
         ]);
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage(
-            'SomeEnum values must be an array with value names as keys.'
-        );
+
+        $this->expectExceptionObject(new InvariantViolation(
+            'SomeEnum values must be an array with value names as keys or values.'
+        ));
         $enumType->assertValid();
     }
 
@@ -2045,5 +2042,28 @@ class DefinitionTest extends TestCase
         );
 
         $objType->assertValid();
+    }
+
+    public function testReturningFieldsUsingYield()
+    {
+        $type = new ObjectType([
+            'name'   => 'Query',
+            'fields' => static function (): Generator {
+                yield 'url'    => ['type' => Type::string()];
+                yield 'width'  => ['type' => Type::int()];
+            },
+        ]);
+
+        $blogSchema = new Schema(['query' => $type]);
+
+        self::assertSame($blogSchema->getQueryType(), $type);
+
+        $field = $type->getField('url');
+        self::assertSame($field->name, 'url');
+        self::assertInstanceOf(StringType::class, $field->getType());
+
+        $field = $type->getField('width');
+        self::assertSame($field->name, 'width');
+        self::assertInstanceOf(IntType::class, $field->getType());
     }
 }